+++ wok_code/wok_code/gen_addons_table.py
@@ -34,7 +34,7 @@ standard_library.install_aliases()  # noqa: E402
 
 MARKERS = r"(\[//\]: # \(addons\))|(\[//\]: # \(end addons\))"
 MANIFESTS = ("__openerp__.py", "__manifest__.py")
-__version__ = "2.1.0"
+__version__ = "2.1.1"
 
 
 class UserError(Exception):